requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
defusedxml==0.7.1
Django==5.2.6
django-cors-headers==4.8.0
django-environ==0.12.0
django-multiselectfield==1.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
djoser==2.3.3
idna==3.10
inflection==0.5.1
Markdown==3.9
oauthlib==3.3.1
pillow==11.3.0
pycparser==2.23
PyJWT==2.10.1
python3-openid==3.2.0
PyYAML==6.0.2
requests==2.32.5
requests-oauthlib==2.0.0
setuptools==80.9.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
wheel==0.45.1
celery>=5.0
django-celery-beat>=2.0
redis==4.2.2
